allow to specify also building category
ONI (since some time later than this mod was started?) first groups
buildings in the menu by category and only keeps order within each
group. In order to be able to move a building to a specific position
in a menu, it may be necessary to specify a category too (especially
with some modded buildings that do not specify their category,
get the default 'uncategorized' and end up first in the menu).

This commit allows to explicitly specify a category, otherwise
it keeps the existing category (if specified, was buggy with modded
buildings).